Change Reveal event hook to prevent glitches on submitting placeholders
The updateCurrentSlide function is used to trigger focusing on
the input text box. Previously, hooking into the slidechanged event
led to the text box receiving focus before the transition finished,
which led to glitchy visual behavior. Hooking into the slidetransitioned
event now leads to the text box receiving focus after the transition,
which leads to a smoother experience.